backend/requirements.txt
fastapi==0.111.1
uvicorn==0.38.0
pydantic==2.7.1
numpy<2.0.0
opencv-python==4.8.1.78
Pillow>=10.0.0
protobuf<=3.20.2
python-dotenv==1.0.1
groq
openai>=1.30.0
httpx
requests

langchain==0.1.12
langchain-community==0.0.29
langchain-huggingface
langgraph
langchain-openai==0.1.0
chromadb==0.4.24
pypdf==4.2.0
sentence-transformers==2.7.0
